flex-grow: determines how the leftover space in the container should be split among the elements of the container. it defines the ratio of how much each elements should take up of the leftover space.
example: in this case, flex grow is set to .4 for both elements of the container. this means, that after both of them taking up 100-100px for the flex basis, each of them get 40% of the leftover space. but because of this, they only take up 80% of the leftover space, and there will still be available space. if we set flex grow to .5 or anything above that, it will fill the container out, and won't overflow.